# services/store-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Store Core API",
        "description": "Этапы 1-2: multi-item orders, partial/refund, untrusted supplier, admin reconcile/recover, rate-limit progress, point-in-time.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local Docker"
        }
    ],
    "paths": {
...
        "/api/v1/admin/finance/period": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Finance period totals",
                "description": "Итоги financial_ledger за период (задача 4).",
                "operationId": "adminFinancePeriod",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Итоги периода",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FinancePeriodResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "AdminToken": []
                    }
                ]
            }
        },
...
        "/api/v1/catalog/stock": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Витрина: SKU и остатки",
                "description": "Пагинированный список активных SKU с остатком из product_stock_cache.",
                "operationId": "catalogStock",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 100,
                            "default": 100,
                            "maximum": 200,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список SKU",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CatalogStockResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
...
        "/api/v1/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Создать заказ",
                "description": "Создаёт заказ: items[{sku,qty}] или legacy sku. Pending webhook до создания заказа применится автоматически.",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "SKU не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получить заказ",
                "description": "Возвращает заказ: header status, amount, items[] (status/issued_code/supplier per line).",
                "operationId": "getOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Публичный ID заказа",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ord_abc123xyz0"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateOrderRequest": {
                "properties": {
                    "sku": {
                        "description": "SKU одного товара (legacy). Либо sku, либо items.",
                        "type": "string",
                        "example": "KEY-CS2-PRIME"
                    },
                    "items": {
                        "description": "Позиции заказа (задача 1 этапа 2: несколько товаров)",
                        "type": "array",
                        "items": {
                            "required": [
                                "sku"
                            ],
                            "properties": {
                                "sku": {
                                    "type": "string",
                                    "example": "KEY-CS2-PRIME"
                                },
                                "qty": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    },
                    "public_id": {
                        "type": "string",
                        "example": "ord_abc123xyz0",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "sku": {
                        "type": "string",
                        "example": "KEY-CS2-PRIME"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 1
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1290
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 1290
                    },
                    "currency": {
                        "type": "string",
                        "example": "RUB"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "issued_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "supplier": {
                        "type": "string",
                        "example": "primary",
                        "nullable": true
                    },
                    "delivered_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "refunded_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "ord_abc123xyz0"
                    },
                    "sku": {
                        "type": "string",
                        "example": "KEY-CS2-PRIME"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "currency": {
                        "type": "string",
                        "example": "RUB"
                    },
                    "status": {
                        "description": "created|paid|delivering|delivered|partially_delivered|refunded|payment_failed|...",
                        "type": "string",
                        "example": "partially_delivered"
                    },
                    "issued_code": {
                        "description": "Legacy single-item; для multi-item смотри items[].issued_code",
                        "type": "string",
                        "example": "CODE-123",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "delivered_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Order"
                    }
                },
                "type": "object"
            },
            "ErrorMessage": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "SKU not found"
                    }
                },
                "type": "object"
            },
...
            "CatalogStockItem": {
                "properties": {
                    "sku": {
                        "type": "string",
                        "example": "KEY-CS2-PRIME"
                    },
                    "name": {
                        "type": "string",
                        "example": "CS2 Prime"
                    },
                    "type": {
                        "type": "string",
                        "example": "key"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "currency": {
                        "type": "string",
                        "example": "RUB"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 100
                    },
                    "total": {
                        "type": "integer",
                        "example": 5000
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "CatalogStockResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CatalogStockItem"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
...
            "FinancePeriodResponse": {
                "properties": {
                    "from": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "to": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "payment_received": {
                        "type": "number"
                    },
                    "delivery_completed": {
                        "type": "number"
                    },
                    "refund_issued": {
                        "type": "number"
                    },
                    "net": {
                        "type": "number"
                    },
                    "counts": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "AdminToken": {
                "type": "apiKey",
                "description": "Admin token. По умолчанию: dev-admin-token",
                "name": "X-Admin-Token",
                "in": "header"
            }
        }
    },
...
}